requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
exceptiongroup==1.2.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
pillow==10.4.0
platformdirs==4.3.6
requests==2.32.3
requests-cache==1.2.1
six==1.16.0
typing-extensions==4.12.2
url-normalize==1.4.3
urllib3==2.2.3
yarl==1.13.1
